requirements.txt
pyspark~=3.2.0
numpy~=1.22.2
py4j~=0.10.9
scipy~=1.8.0
pandas~=1.4.1
setuptools~=56.0.0
pip~=21.1.1
lxml~=4.8.0
tornado~=6.1
Cython~=0.29.23
nltk~=3.7
flair~=0.10
matplotlib~=3.5.1
sklearn~=0.0
scikit-learn~=1.0.2
zeep~=4.1.0
requests~=2.27.1
bs4~=0.0.1
beautifulsoup4~=4.10.0

pytorch-pretrained-bert==0.6.2
numpy==1.16.4
torch==1.2.0
torchvision==0.4.0
scikit-learn==0.20.3
pandas==0.23.0